Provide Update Config script that will allow users to tweak specific settings
It's a problem on the side of users to be continuously add/remove write
permissions from some of the .ini files.

This script removes that issue entirely, resulting in an isolated
environment that allows users to change things at their will and then
running this script afterwards to update their game config.